requirements.txt
websocket-client==1.0.1
requests==2.22.0
